TASK: Rename Doctrine\Common\Persistence -> Doctrine\Persistence
